ENH: Make scattertext() a public API
This will make it easier to use for various use cases.
